+++ switchstate/models/_IFHMM.py
@@ -1,7 +1,7 @@
-import time
 import torch
-from tqdm.auto import tqdm
-from ..utils import log_domain_mean
+
+from .methods import log_transform_params
+from .trainer import train
 
 class IFHMM(torch.nn.Module):
     
@@ -65,17 +65,7 @@ class IFHMM(torch.nn.Module):
 
     def forward_model(self):
     
-        # Normalise across states 
-        self.log_state_init = torch.nn.functional.log_softmax(self.unnormalized_state_init, dim=0)
-
-        # Normalise chain weights
-        self.log_chain_weights = torch.nn.functional.log_softmax(self.unnormalized_chain_weights, dim=-1)
-        
-        # Normalise across chains for each state
-        self.log_emission_matrix = torch.nn.functional.log_softmax(self.unnormalized_emission_matrix, dim=-1)
-
-        # Normalise TPM so that they are probabilities (log space)
-        self.log_transition_matrix = torch.nn.functional.log_softmax(self.unnormalized_transition_matrix, dim=-1)
+        log_transform_params(self)
 
         # MSM iteration wise probability calculation
         log_hidden_state_probs = torch.zeros(self.num_iters, self.num_states, self.num_chains)
@@ -105,6 +95,8 @@ class IFHMM(torch.nn.Module):
     # Conditional hidden state probabilities given observed sequence till current iteration
     def filtering(self, D):
 
+        log_transform_params(self)
+
         log_alpha = torch.zeros(self.num_iters, self.num_states, self.num_chains)
         log_probs = torch.zeros(self.num_iters, self.num_chains)
         if self.is_cuda: 
@@ -128,6 +120,8 @@ class IFHMM(torch.nn.Module):
 
     # Conditional hidden state probabilities given an observed sequence
     def smoothing(self, D):
+
+        log_transform_params(self)
 
         log_alpha, log_probs = self.filtering(D)
         log_beta = torch.zeros(self.num_iters, self.num_states, self.num_chains)
@@ -155,6 +149,8 @@ class IFHMM(torch.nn.Module):
         
     # Viterbi decoding for best path
     def viterbi(self, D):
+
+        log_transform_params(self)
 
         log_delta = torch.zeros(self.num_iters, self.num_states, self.num_chains)
         psi = torch.zeros(self.num_iters, self.num_states, self.num_chains)
@@ -187,134 +183,12 @@ class IFHMM(torch.nn.Module):
             best_path.append(best_path_i)
 
         return log_delta, psi, log_max, best_path
-
-    def train(self, D, TPM=None, num_epochs=300, sparsity_weight=1.0, 
+    
+    # Train the model
+    def train(self, D, TPM=None, num_epochs=300, sparsity_weight=1.0,
               optimizer=None, criterion=None, swa_scheduler=None, swa_start=200, 
               verbose=False):
-        
-        '''
-        Train the model.
-        
-        Parameters
-        ----------
-        D : FloatTensor of shape (MSM_nodes, T_max)
-            MSM simulation data.
-        TPM : Transition probability matrix
-            Used to regularise emission probabilities.
-        num_epochs : int (default: 300)
-            Number of training epochs
-        sparsity_weight : float (default: 1.0)
-            Regularisation weight for sparse latent TPM.
-        optimizer : (default: RMSprop(lr=0.2))
-            Optimizer algorithm.
-        criterion : (default: KLDivLoss())
-            Loss function. (Default preferred)
-        swa_scheduler : (default: None)
-            SWA scheduler.
-        swa_start : int (default: 200)
-            Training epoch to start SWA.
-        verbose : bool (default: False)
-            Toggle printing of training history
-        
-        '''
-
-        if self.is_cuda:
-            D = D.cuda()
-        
-        self.sparsity_weight = sparsity_weight
-         
-        try: assert self.elapsed_epochs
-        except: self.elapsed_epochs = 0
-        
-        try: assert self.loss_values
-        except: self.loss_values = []
-
-        try: assert self.divergence_values
-        except: self.divergence_values = []
-            
-        try: assert self.likelihood_values
-        except: self.likelihood_values = []
-
-        try: assert self.sparsity_values
-        except: self.sparsity_values = []
-
-        try: assert self.regularisation_values
-        except: self.regularisation_values = []
-
-        try: assert self.independence_values
-        except: self.independence_values = []
-        
-        self.optimizer = optimizer
-        if self.optimizer is None:
-            self.optimizer = torch.optim.RMSprop(self.parameters(), lr=0.2)
-        self.swa_scheduler = swa_scheduler
-        self.criterion = criterion
-        if self.criterion is None:
-            self.criterion = torch.nn.KLDivLoss(reduction='batchmean', log_target=False)
-                
-        start_time = time.time()           
-        for epoch in tqdm(range(num_epochs), desc='Training dynamic model'):
-
-            self.optimizer.zero_grad()
-            prediction, log_observed_state_probs_, log_hidden_state_probs = self.forward_model()
-            log_alpha, log_probs = self.filtering(D)
-            
-            divergence = self.criterion(prediction, D)
-            loss = divergence
-            self.divergence_values.append(divergence.item())
-
-            likelihood = log_alpha.max(1)[0].logsumexp(0).logsumexp(0)
-            #loss = -likelihood
-            self.likelihood_values.append(likelihood.item())
-
-            sparsity = self.criterion(self.log_transition_matrix, torch.eye(self.num_states))
-            loss += self.sparsity_weight*sparsity
-            self.sparsity_values.append(sparsity.item())
-
-            independence = log_domain_mean(log_domain_mean(log_observed_state_probs_).logsumexp(0).sum(0)).squeeze()
-            loss -= 1/independence
-            self.independence_values.append(independence.item())
-
-            if TPM is not None:
-                regularisation =  self.criterion(self.log_emission_matrix,
-                                                 torch.matmul(torch.exp(self.log_emission_matrix.detach()), TPM))
-                loss += regularisation
-                self.regularisation_values.append(regularisation.item())
-
-            loss.backward()
-            self.optimizer.step()
-            if self.elapsed_epochs > swa_start and self.swa_scheduler is not None:
-                swa_scheduler.step()
-            
-            self.loss_values.append(loss.item())
-
-            self.elapsed_epochs += 1
-            if epoch % 10 == 0 or epoch <=10:
-                corrcoeffs = []
-                outputs = torch.exp(prediction)
-                for t in range(self.num_iters):
-                    corrcoeffs.append(torch.corrcoef(torch.stack([outputs[t], D[t]]))[1,0])
-                avg_corrcoeff = torch.mean(torch.tensor(corrcoeffs))
-
-                # Print Loss
-                if verbose:
-                    if TPM is not None:
-                        print('Time: {:.2f}s. Iter: {}. Loss: {:.2E}. KL: {:.2E}. Likl: {:.2E}. Sparse: {:.2E}. Reg: {:.2E}. Ind: {:.2E}. Corcoef: {:.2f}.'.format(time.time() - start_time,
-                                                                                    self.elapsed_epochs, 
-                                                                                    self.loss_values[-1],
-                                                                                    self.divergence_values[-1],
-                                                                                    self.likelihood_values[-1],
-                                                                                    self.sparsity_values[-1],
-                                                                                    self.regularisation_values[-1],
-                                                                                    self.independence_values[-1],
-                                                                                    avg_corrcoeff))
-                    else:
-                        print('Time: {:.2f}s. Iter: {}. Loss: {:.2E}. KL: {:.2E}. Likl: {:.2E}. Sparse: {:.2E}. Ind: {:.2E}. Corcoef: {:.2f}.'.format(time.time() - start_time,
-                                                                                    self.elapsed_epochs, 
-                                                                                    self.loss_values[-1],
-                                                                                    self.divergence_values[-1],
-                                                                                    self.likelihood_values[-1],                                                    
-                                                                                    self.sparsity_values[-1],
-                                                                                    self.independence_values[-1],
-                                                                                    avg_corrcoeff))
+        train(self, D, TPM=TPM, num_epochs=num_epochs, sparsity_weight=sparsity_weight,
+            optimizer=optimizer, criterion=criterion, swa_scheduler=swa_scheduler, swa_start=swa_start, 
+            verbose=verbose)
