+++ switchstate/models/__init__.py
@@ -1,3 +1,3 @@
-from .IFHMM import IFHMM
-from .SSM import SSM
-from .SSM_node import SSM_node
+from ._IFHMM import IFHMM
+from ._SSM import SSM
+from ._SSM_node import SSM_node
